# [WIP] recursive, depth-first optimizer