… docs/source/api.rst
… docs/source/branching_path.rst
… docs/source/eager_path.rst
… docs/source/greedy_path.rst
… docs/source/index.rst
… docs/source/optimal_path.rst
… docs/source/path_finding.rst
… docs/source/path_finding_time.png
… docs/source/path_found_flops.png
… docs/source/reusing_paths.rst
… opt_einsum/tests/test_paths.py