## Use recursive, depth-first optimizer for optimal & possibly 'greedy2'

I had a play around with making the ``opt_einsum`` optimal path finder a bit more efficient. The basic idea is:

1. Use a recursive search to simplify the logic and find paths 'depth-first', i.e. follow a path all the way to its termination before searching for the next one.
2. Allowed by this, terminate searches as soon as the path cost increases above the best found so far, sieving many known-bad paths

The results are pretty promising, current optimal:

```python
>> eq, shapes = oe.helpers.rand_equation(8, 3, seed=42)
>> views = [np.random.rand(*s) for s in shapes]
>> %time path = oe.contract_path(eq, *views, optimize='optimal')[1]
CPU times: user 25.3 s, sys: 716 ms, total: 26 s
Wall time: 26 s
>>> path.opt_cost
130316
```

and with this recursive strategy:

```python
>> %time path = oe.contract_path(eq, *views, optimize='roptimal')[1]
CPU times: user 107 ms, sys: 324 µs, total: 107 ms
Wall time: 108 ms
>>> path.opt_cost
130316
```

The fundamental scaling is still factorial, but it should allow optimal-paths to be realistically found for expressions of up to 10 tensors. Possibly other heuristics can be used to sieve paths eagerly as well.

One additional option this opens up is to recurse only into the ``n`` best possible contractions (current greedy/eager being ``n=1``). Thus you could have a ``'greedy2'``, ``'greedy3'`` etc, which scale like ``2**n``, ``3**n`` which while exponential (modulo sieving), is significantly more manageable than factorial scaling. These might use custom cost functions #64 and random shaking as well #63.

Anyway, thought I'd share this WIP initial implementation. My suggestion might be to have this approach eventually replace optimal, with ``'greedy'``, ``'greedy2'``... using it also.

cc @fritzo --- took some inspiration from your ``'eager'`` path if you have any input!

TODO:
---------
- [x] apply some caching?
- [ ] keep track of ``keep`` indices rather than computing each time?
- [x] handle memory_limit case properly? (i.e. contract all remaining tensors at once)