@dgasmith 2018-10-21
This is pretty cool and a good way to trim a lot of bad paths early. You are right that it provides a lot of interesting paths to explore in terms of different cost functions and path randomness.

Something to try would be to evaluate a greedy path at the very start and use that as a sieving upper bound.