@jcmgray 2018-10-21
> Something to try would be to evaluate a greedy path at the very start and use that as a sieving upper bound.

Yes definitely, so if you sort the contractions/use a heap at each step (as you might before recursing into only the best ``N``) , the very first path that is analysed will be the greedy path, setting a reasonable sieve.

--- 

Regarding caching, am I right in thinking that for a given ``k1``, ``k2`` pair, regardless of what other contractions have happened, the resulting ``k12`` and flop cost will be the same? This would be a nice saving. It's definitely true for 'standard' contractions where indices appear no more than twice, and think so otherwise but not totally sure, (@dgasmith / @fritzo)?