@dgasmith 2018-10-22
`abc, abc -> abc` vs `abc, abc ->`. The cost of the latter requires an extra factor of two (unless you use FMA which you probably do...) as you need an extra addition per element.