@jcmgray 2018-10-22
> abc, abc -> abc vs abc, abc ->. The cost of the latter requires an extra factor of two (unless you use FMA which you probably do...) as you need an extra addition per element.

Aha while you are right here, these two will actually never appear within the same contraction as the output is constant throughout. If indices should be kept as they appear on other inputs

So I've tentatively enabled caching and it seems to produce the correct results so far (and reduces the time for the path finding above to ``~25ms``.