@dgasmith 2018-10-23
With caching we might need to watch out how much we are caching here. The factorial compute is associated with factorial memory requirements. The speedups are quite good however with caching!