@jcmgray 2018-10-23
I've updated the optimal approach a bit and implemented a recursive greedy strategy (orientated around greedily taking *multiple* paths at once rather than just a single one).

Some broadlly illustrative numbers using the setup above:

## new optimal 

(can realistically handle up to about ~10 tensors):
```python
>>> oe.contract_path(eq, *views, optimize='roptimal')[1].opt_cost
130316

>>> %timeit oe.contract_path(eq, *views, optimize='roptimal')[1]
15.6 ms ± 273 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```

## 'greedy-all' / 'branch-all'

Quite similar to optimal but only assess inner products, and actively terminate paths if they relatively become bad rather than waiting (can realistically handle up to about ~14 tensors):

```python
>>> oe.contract_path(eq, *views, optimize='rgreedy')[1].opt_cost
130362

>>> %timeit oe.contract_path(eq, *views, optimize='rgreedy')[1]
1.26 ms ± 17.2 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

## 'greedy2' / 'branch-2'

As above, but only keep the two best contractions upon each recurse (can realistically handle up to about ~20 tensors):

```python
>>> oe.contract_path(eq, *views, optimize='rgreedy2')[1].opt_cost
130362

>>> %timeit oe.contract_path(eq, *views, optimize='rgreedy2')[1]
757 µs ± 3.45 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

And then the old greedy and eager for comparison:

```python
>>> # greedy
>>> oe.contract_path(eq, *views, optimize='greedy')[1].opt_cost
520152

>>> %timeit oe.contract_path(eq, *views, optimize='greedy')[1]
423 µs ± 4.13 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

>>> # eager
>>> oe.contract_path(eq, *views, optimize='eager')[1].opt_cost
815064

>>> %timeit oe.contract_path(eq, *views, optimize='eager')[1]
429 µs ± 3.08 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

So the new recursive greedy (which maybe needs a better name to differentiate? ``'branch-all'``, ``branch-2``?) works quite well as an intermediate point between quality and run-time.

I imagine now:
1. Just replacing the ``optimal`` path, the new one I think finds the exact same paths.
2. Adding ``'branch-all'``, ``'branch-2'`` as options.
3. Removing the old ``greedy`` implementation and making it point to ``eager`` which is faster while finding roughly the same quality of contractions
4. Tweaking ``'auto'`` to dispatch to optimal, then the branching/recursive greedy, then eager.