@dgasmith 2018-10-23
Can you check to make sure the new and old optimal all give the same results by trying maybe 1-10k random paths for 1-6 tensors? If that passes I think we should actually look at removing the current optimal path, I believe this does exactly the same thing and will have much better performance characteristics due to more aggressive pruning.

I like the `branch` name for the new category of optimizers as I don't quite think the characteristics are similar to greedy.

It doesn't hurt to have many different paths, but the rate we are growing is pretty steep and it is not clear if the various paths are "better" than another and what "better" really means. We will have to sum this up to users in a succinct way at some point and make recommendations. Work is extremely busy at the moment so I haven't had a chance to dive into this.