@jcmgray 2018-10-24
> Can you check to make sure the new and old optimal all give the same results by trying maybe 1-10k random paths for 1-6 tensors?

I've now done this for various contractions from ``rand_equation`` and the flop cost found is always identical (though occasionally a different path is found - an 'equally optimal' one). I might add 3+ index contractions to be very certain at some point later.