@jcmgray 2018-10-30
Ok sorry for the delay I've updated this PR a bit:

1. made the recursive strategy the default optimal algorithm
2. made ``'greedy'`` point to the eager algorithm
3. added ``'branch-all'`` and ``'branch-2'`` as options
4. ``'auto'`` mode now uses :
    -  ``'optimal'`` for <= 4 inputs
    -  ``'branch-all'`` for <= 6 inputs
    -  ``'branch-2'`` for <= 8 inputs
    - ``'greedy'`` / ``'eager'`` otherwise.

   These cutoffs are chosen such that the path finding is always well below ``1ms``, but with nice increases in quality of contraction for up to 8 tensors.

Currently I've made ``'greedy'`` with a ``memory_limit`` set point to ``branch(..., nbranch=1)``, which does essentially the same thing but quite a lot less efficiently. The options here are:
* a) leave it like this
* b) just ignore ``memory_limit`` for greedy contractions (maybe with a warning if set?)
* c) modify ``eager`` to accept a ``memory_limit`` (most work and might make ``eager`` slower)

## TODO:
- [ ] Decide on a ``memory_limit`` strategy for greedy/eager
- [ ] Document the changes
- [ ] Make ``branch`` more customizable? I would envisage something like 

```python
contract(..., optimize=oe.BranchOptimizer(nbranch=3, flops_cutoff=1.5))
```
with a similar syntax maybe for #63. I suspect ``branch`` will mostly be used as part of the ``auto`` strategy so this might not be that important an addition.